loader: create Loader type
This type is a wrapper around operations that previously had no receiver, as
well as on fields which were package-level variables. These variables are now
all contained in the Loader type. As a result of making the Loader type,
we can declare an interface in `pkg/datapath` which Loader implements. The
functionality of `pkg/loader` is now hidden behind this interface.

The Datapath interface has been extended to have a `Loader()` function which
returns the Loader implementation of whatever the type implementing the
Datapath interface desires.

As part of this rework, the configuration-writing pieces of the Datapath
interfac have been factored out into their own interface, called ConfigWriter.
This interface is now used by `pkg/loader` instead of the Datapath interface to
avoid cyclic import cycles in testing.

Signed-off by: Ian Vernon <ian@cilium.io>